Fixes bug related to restoring training
