+++ l2hmc-qcd/train.py
@@ -105,167 +105,12 @@ def dict_to_str(d):
     return '\n'.join(strs)
 
 
-def load_configs_from_logdir(logdir: Union[str, Path]):
-    fpath = os.path.join(logdir, 'train_configs.json')
-    with open(fpath, 'r') as f:
-        configs = json.load(f)
-
-    return configs
-
-
-def setup(configs: dict[str, Any]):
-    """Setup for training."""
-    # Create copy of configs.
-    # Try loading configs from `configs.logdir/train_configs.json` and restore
-    if hasattr(configs, '__dict__'):
-        configs = getattr(configs, '__dict__')
-
-    output = deepcopy(configs)
-    logfile = os.path.join(os.getcwd(), 'log_dirs.txt')
-
-    ensure_new = configs.get('ensure_new', False)
-    logdir = configs.get('logdir', configs.get('log_dir', None))
-    if logdir is not None:
-        logdir_exists = os.path.isdir(logdir)
-        contents = os.listdir(logdir)
-        logdir_nonempty = False
-        if contents is not None and isinstance(contents, list):
-            if len(contents) > 0:
-                logdir_nonempty = True
-
-        if logdir_exists and logdir_nonempty and ensure_new:
-            raise ValueError(
-                f'Nonempty `logdir`, but `ensure_new={ensure_new}'
-            )
-
-    dirs = io.setup_directories(configs, timestamps = TSTAMPS)
-    if ensure_new:
-        dirs = io.setup_directories(configs, timestamps=TSTAMPS)
-        output['dirs'] = dirs
-        #  logdir = io.make_log_dir(configs, 'GaugeModel', logfile)
-        output['log_dir'] = logdir
-        output['logdir'] = logdir
-        output['restore'] = False
-        #  io.write(f'{logdir}', logfile, 'a')
-
-    else:
-        dirs = io.setup_directories(configs, timestamps=TSTAMPS)
-        logdir = dirs['logdir']
-        output['restore'] = True
-        output['logdir'] = logdir
-        output['log_dir'] = logdir
-        output['restored_from'] = logdir
-        if logdir is not None:
-            output = load_configs_from_logdir(logdir)
-            output['restored'] = True
-            output['log_dir'] = logdir
-            output['logdir'] = logdir
-            output['restored_from'] = logdir
-            to_overwrite = ['train_steps', 'run_steps', 'beta_final']
-            changed = {k: configs.get(k) for k in to_overwrite}
-            for key, new in changed.items():
-                old = output.get(key)
-                if new != old:
-                    logger.warning(
-                        f'Overwriting {key} from {old} to {new} in configs'
-                    )
-                    output[key] = new
-
-    num_chains = output.get('num_chains', 16)
-    if output.get('hmc_steps', 0) > 0:
-        x, hdynamics, _, hflags = train_hmc(output, num_chains=num_chains)
-        _ = run(hdynamics, hflags, save_x=False)
-
-    if RANK == 0:
-        io.write(f'{logdir}', logfile, 'a')
-
-    return output
-
-
 def main(configs: dict[str, Any]):
     """Main method for training."""
     # TODO: Move setup code to separate function and refactor
-    configs = setup(configs)
+    #  configs = setup(configs)
     #  tf.keras.backend.set_floatx('float32')
-
-    x = None
-    logdir = configs.get('log_dir', None)
-    #  beta_init = configs.get('beta_init', None)
-    #  beta_final = configs.get('beta_final', None)
-    #  ensure_new = configs.get('ensure_new', False)
-    #  logfile = os.path.join(os.getcwd(), 'log_dirs.txt')
-
-    #  if logdir is None:
-    #      # Check and see if `./log_dirs.txt` exists and if so, try loading from
-    #      # there if configs are compatible
-    #      # TODO: Check if configs are explicitly compatible
-    #      if os.path.isfile(logfile):
-    #          logger.rule('Found `log_dirs.txt`!')
-    #          logger.log(f'logfile: {logfile}')
-    #          logdirs = []
-    #          with open(logfile, 'r') as f:
-    #              for line in f.readlines():
-    #                  logdirs.append(line.rstrip('\n'))
-    #
-    #          logdir = logdirs[-1]
-    #          if os.path.isdir(logdir):
-    #              if ensure_new:
-    #                  err = (f'configs["ensure_new"] = {ensure_new}, but '
-    #                         f'found existing `log_dir` at:\n'
-    #                         f'  log_dir: {logdir}. Exiting!!')
-    #                  raise ValueError(err)
-    #
-    #              else:
-    #                  logdir = None
-    #
-    #          # If `./log_dirs.txt` exists and points to a directory containing
-    #          # training checkpoints
-    #          #  if ensure_new and os.path.isdir(logdir):
-    #          #      ckptdir = os.path.join(logdir, 'training', 'checkpoints')
-    #          #      exists = (len(os.listdir(ckptdir)) > 0)
-    #          #      if exists:
-    #          #          err = (f'configs["ensure_new"] = {ensure_new}, but '
-    #          #                 f'found checkpoints in `log_dir` from '
-    #          #                 f'./`log_dirs.txt. Exiting!\n'
-    #          #                 f'  log_dir: {logdir}')
-    #          #          raise ValueError(err)
-    #          #
-    #          #      #  exist = len(os.listdir(ckptdir)) > 0 if os.path.isdir(ckptdir)
-    #          #      #  if len(ckpts) > 0:
-    #          #      #      raise
-    #          #      #  if os.path.isdir(ckptdir):
-    #          #      #      ckpts = os.listdir(ckptdir)
-    #          #      #      if len(ckpts) > 0:
-    #          #      #          err = (f'configs["ensure_new"] = {ensure_new}, but '
-    #          #      #                 f'found checkpoints in `log_dir` from '
-    #          #      #                 f'./`log_dirs.txt. Exiting!\n'
-    #          #      #                 f'  log_dir: {logdir}')
-    #          #      #          raise ValueError(err)
-
-    #  if logdir is not None:  # we want to restore from latest checkpoint
-    #      #  logdir = configs.get('log_dir', None)
-    #      run_steps = configs.get('run_steps', None)
-    #      train_steps = configs.get('train_steps', None)
-    #      restored = restore_flags(configs, os.path.join(logdir, 'training'))
-    #      for key, val in configs.items():
-    #          if key in restored:
-    #              if val != restored[key]:
-    #                  io.log(f'Restored {key}: {restored[key]}')
-    #                  io.log(f'Using {key}: {val}')
-    #
-    #      configs['run_steps'] = run_steps
-    #      configs['train_steps'] = train_steps
-    #      if beta_init != configs.get('beta_init', None):
-    #          configs['beta_init'] = beta_init
-    #      if beta_final != configs.get('beta_final', None):
-    #          configs['beta_final'] = beta_final
-    #
-    #      configs['restore'] = True
-    #      configs['restored_from'] = logdir
-
-    #  else:  # New training session
-    #  x, dynamics, train_data, configs = train(configs=configs,
-    train_out = train(configs=configs, x=x, make_plots=True) # , num_chains=nchains)
+    train_out = train(configs=configs, make_plots=True) # , num_chains=nchains)
     x = train_out.x
     logdir = train_out.logdir
     #  train_data = train_out.data
@@ -277,8 +122,6 @@ def main(configs: dict[str, Any]):
     # ====
     # Run inference on trained model
     if run_steps > 0:
-        #  run_steps = args.get('run_steps', 125000)
-        #  logdir = configs.log_dir
         beta = configs.get('beta_final')
         nchains = configs.get('num_chains', configs.get('nchains', 16))
         if configs.get('small_batch', False):
